refactor: close legacy command boundaries
Move command-only MerkleDAG import support under cmd/internal, remove legacy resolve transcript DTO/server helpers, and delete stale UnixFS arc counters.

Add source-layout guards for those boundaries.
