+++ cmd/eval/helper/evalwrite/command.go
@@ -15,7 +15,7 @@ import (
 	gittrace "github.com/dewebprotocol/malt/cmd/eval/helper/git"
 	"github.com/dewebprotocol/malt/cmd/eval/helper/replay"
 	evalstore "github.com/dewebprotocol/malt/cmd/eval/helper/store"
-	"github.com/dewebprotocol/malt/internal/merkledagimport"
+	"github.com/dewebprotocol/malt/cmd/internal/merkledagimport"
 	"github.com/spf13/cobra"
 )
