+++ cmd/eval/internal/eval/readbench/baseline.go
@@ -9,9 +9,9 @@ import (
 	"strings"
 	"time"
 
+	"github.com/dewebprotocol/malt/cmd/internal/merkledagimport"
 	casmock "github.com/dewebprotocol/malt/core/cas/mock"
 	"github.com/dewebprotocol/malt/core/metrics"
-	"github.com/dewebprotocol/malt/internal/merkledagimport"
 	unixfsio "github.com/ipfs/boxo/ipld/unixfs/io"
 	cid "github.com/ipfs/go-cid"
 	ipld "github.com/ipfs/go-ipld-format"
